requirements.txt
Flask
sqlalchemy
psycopg2-binary
gunicorn